010--Rename some model files Show remaining labels at summary in polish language
