+++ src/app/pages/order/order.component.ts
@@ -14,8 +14,8 @@ import { MessageService } from 'primeng/api';
 import { FloatLabelModule } from 'primeng/floatlabel';
 import { Store } from '@ngrx/store';
 import { storeOrderDetails } from '../../store/order.actions';
-import { OrderDetails } from '../../shared/models/purchase-data.model';
-import { Order } from '../../shared/models/order-details.model';
+import { OrderDetails } from '../../shared/models/order-details.model';
+import { Order } from '../../shared/models/order.model';
 import { Observable, take } from 'rxjs';
 import { DividerModule } from 'primeng/divider';
 import { FormErrorComponent } from '../../shared/form-error/form-error.component';
