+++ src/app/pages/order/recipient-form/recipient-form.component.ts
@@ -5,7 +5,7 @@ import { InputTextModule } from 'primeng/inputtext';
 import { FloatLabelModule } from 'primeng/floatlabel';
 import { RECIPIENT_FORM } from './recipient-form.data';
 import { TypedForm } from '../../../shared/models/typed-form.model';
-import { RecipientData } from '../../../shared/models/purchase-data.model';
+import { RecipientData } from '../../../shared/models/order-details.model';
 import { ButtonModule } from 'primeng/button';
 import { FormErrorComponent } from '../../../shared/form-error/form-error.component';
 import { DynamicFormPipe } from './dynamic-form.pipe';
@@ -20,7 +20,7 @@ import { DynamicFormPipe } from './dynamic-form.pipe';
     FloatLabelModule,
     ButtonModule,
     FormErrorComponent,
-    DynamicFormPipe
+    DynamicFormPipe,
   ],
   templateUrl: './recipient-form.component.html',
   styleUrl: './recipient-form.component.scss',
@@ -34,6 +34,6 @@ export class RecipientFormComponent {
   readonly recipientFormData = RECIPIENT_FORM;
 
   removeRecipient(id: number) {
-    this.removeRecipientOutput.emit(id)
+    this.removeRecipientOutput.emit(id);
   }
 }
