+++ src/app/pages/summary/summary.component.html
@@ -18,7 +18,7 @@
               Rodzaj płatności:
             </span>
             <span>
-              {{orderDetails.paymentType}}
+              {{orderDetails.paymentType | getLabel : paymentTypes}}
             </span>
           </div>
           <div class="summary__property">
@@ -26,7 +26,7 @@
               Rodzaj dostawy:
             </span>
             <span>
-              {{orderDetails.deliveryType}}
+              {{orderDetails.deliveryType | getLabel : deliveryTypes}}
             </span>
           </div>
           @if (orderDetails.deliveryDate) {
