+++ src/app/pages/summary/summary.component.ts
@@ -5,10 +5,13 @@ import { AsyncPipe, DatePipe } from '@angular/common';
 import { LawnmowerDetailsComponent } from '../../shared/lawnmower-details/lawnmower-details.component';
 import { Router, RouterLink } from '@angular/router';
 import { UserDataComponent } from './user-data/user-data.component';
-import { Order } from '../../shared/models/order-details.model';
+import { Order } from '../../shared/models/order.model';
 import { ButtonModule } from 'primeng/button';
 import { MessageService } from 'primeng/api';
 import { DataTileComponent } from '../../shared/data-tile/data-tile.component';
+import { PAYMENT_TYPES } from '../order/data/payment-types.data';
+import { DELIVERY_TYPES } from '../order/data/delivery-types.data';
+import { GetLabelPipe } from './get-label.pipe';
 
 @Component({
   selector: 'app-summary',
@@ -20,13 +23,16 @@ import { DataTileComponent } from '../../shared/data-tile/data-tile.component';
     UserDataComponent,
     ButtonModule,
     RouterLink,
-    DataTileComponent
+    DataTileComponent,
+    GetLabelPipe
   ],
   templateUrl: './summary.component.html',
   styleUrl: './summary.component.scss',
 })
 export class SummaryComponent implements OnInit {
   readonly orderData$: Observable<Order>;
+  readonly paymentTypes = PAYMENT_TYPES;
+  readonly deliveryTypes = DELIVERY_TYPES;
 
   constructor(
     private readonly _router: Router,
