+++ src/app/shared/models/order-details.model.ts
@@ -1,8 +1,21 @@
-import { Lawnmower } from './lawnmower.model';
-import { OrderDetails } from './purchase-data.model';
+import { DeliveryType } from '../../pages/order/models/delivery-type.enum';
+import { PaymentType } from '../../pages/order/models/payment-type.enum';
 
-export interface Order {
-  lawnmower: Lawnmower;
-  orderDetails: OrderDetails;
+export interface OrderDetails {
+  paymentType: PaymentType;
+  deliveryType: DeliveryType;
+  deliveryDate: Date;
+  deliveryAddress: RecipientData;
+  recipients: RecipientData[];
 }
 
+export interface RecipientData {
+  firstName: string;
+  lastName: string;
+  city: string;
+  street: string;
+  houseNumber: string;
+  apartmentNumber: string;
+  zipCode: string;
+  id?: number;
+}
