+++ src/app/shared/models/order.model.ts
@@ -1,0 +1,7 @@
+import { Lawnmower } from './lawnmower.model';
+import { OrderDetails } from './order-details.model';
+
+export interface Order {
+  lawnmower: Lawnmower;
+  orderDetails: OrderDetails;
+}
