--- src/app/shared/models/purchase-data.model.ts
@@ -1,21 +1,0 @@
-import { DeliveryType } from '../../pages/order/models/delivery-type.enum';
-import { PaymentType } from '../../pages/order/models/payment-type.enum';
-
-export interface OrderDetails {
-  paymentType: PaymentType;
-  deliveryType: DeliveryType;
-  deliveryDate: Date;
-  deliveryAddress: RecipientData;
-  recipients: RecipientData[];
-}
-
-export interface RecipientData {
-  firstName: string;
-  lastName: string;
-  city: string;
-  street: string;
-  houseNumber: string;
-  apartmentNumber: string;
-  zipCode: string;
-  id?: number;
-}
