+++ src/app/store/order.actions.ts
@@ -1,6 +1,6 @@
 import { createAction, props } from '@ngrx/store';
 import { Lawnmower } from '../shared/models/lawnmower.model';
-import { OrderDetails } from '../shared/models/purchase-data.model';
+import { OrderDetails } from '../shared/models/order-details.model';
 
 export const storeLawnmower = createAction(
   '[Lawnmower] Store Lawnmower',
